moved graph into its own sub-folder
