+++ src/RPGTools/DialogueTreeMaker/DialogueTreeGraph/DialogueTreeGraph.tsx
@@ -3,8 +3,8 @@ import { useLoadGraph, useRegisterEvents, useSigma } from "@react-sigma/core";
 import { SerializedGraph } from "graphology-types";
 import { DirectedGraph } from "graphology";
 
-import { convertDialoguesToEdges, convertDialoguesToNodes } from "./domain/graphUtil";
-import { Dialogue, NodeCoordinate, SkillTest } from "./domain/types";
+import { convertDialoguesToEdges, convertDialoguesToNodes } from "../domain/graphUtil";
+import { Dialogue, NodeCoordinate, SkillTest } from "../domain/types";
 
 type Props = {
     dialogues: Array<Dialogue>;
